+++ examples/nuxt/server/middleware/federation.ts
@@ -1,0 +1,4 @@
+import { integrateFederation } from "@fedify/nuxt";
+import federation from "../federation";
+
+export default integrateFederation(federation, () => undefined);